refactor: moved widget components to `widget/src/components`
