+++ src/components/Footer.tsx
@@ -1,4 +1,4 @@
-import packageJSON from '../../../package.json';
+import packageJSON from '../../package.json';
 
 export const Footer = () => {
   return (
