+++ src/components/InfoBox.tsx
