+++ src/translations/english.json
@@ -42,7 +42,6 @@
   "generator.theme.color_scheme": "Widget theme",
   "generator.theme.style": "Widget style",
   "generator.theme.banner_as_background": "Use profile banner as background",
-  "generator.theme.banner_as_background.warning": "The preview shows an example banner. Correct banner will show after generating the widget.",
   "generator.theme.banner_as_background.no_banner": "Player {{0}} has no profile banner set.",
   "generator.theme.banner_as_background.adjust_opacity": "Adjust banner opacity",
   "generator.generate.title": "Generate",
