+++ widget/src/components/Statistic.tsx
